# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Recall CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Campaigns ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, nome, valor, duracao, ativo]
      properties:
        id:
          type: integer
        nome:
          type: string
        valor:
          type: number
        duracao:
          type: integer
        ativo:
          type: boolean
    ServiceInput:
      type: object
      required: [nome, valor, duracao]
      properties:
        nome:
          type: string
        valor:
          type: number
        duracao:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        nome:
          type: string
        valor:
          type: number
        duracao:
          type: integer
        ativo:
          type: boolean

    # ─── Appointment ────────────────────────────────────────────────
...